+++ core/src/main/java/com/nabla/sdk/core/data/patient/PatientRepositoryImpl.kt
@@ -2,6 +2,7 @@ package com.nabla.sdk.core.data.patient
 
 import com.nabla.sdk.core.domain.boundary.PatientRepository
 import com.nabla.sdk.core.domain.entity.StringId
+import kotlinx.coroutines.flow.Flow
 
 internal class PatientRepositoryImpl(
     private val localPatientDataSource: LocalPatientDataSource,
@@ -13,5 +14,9 @@ internal class PatientRepositoryImpl(
 
     override fun getPatientId(): StringId? {
         return localPatientDataSource.getPatient()
+    }
+
+    override fun getPatientIdFlow(): Flow<StringId?> {
+        return localPatientDataSource.patientIdFlow
     }
 }