+++ core/src/main/java/com/nabla/sdk/core/domain/boundary/SessionClient.kt
@@ -4,9 +4,7 @@ import com.nabla.sdk.core.annotation.NablaInternal
 
 @NablaInternal
 public interface SessionClient {
-    public fun initSession(sessionTokenProvider: SessionTokenProvider)
-    public fun isSessionInitialized(): Boolean
+    public fun authenticatableOrThrow()
     public suspend fun getFreshAccessToken(forceRefreshAccessToken: Boolean = false): String
     public fun markTokensAsInvalid()
-    public fun clearSession()
 }