+++ core/src/main/java/com/nabla/sdk/core/domain/boundary/SessionLocalDataCleaner.kt
@@ -1,5 +1,5 @@
 package com.nabla.sdk.core.domain.boundary
 
 internal interface SessionLocalDataCleaner {
-    fun cleanLocalSessionData()
+    suspend fun cleanLocalSessionData()
 }